+++ src/jwk.rs
@@ -480,6 +480,7 @@ mod helpers {
 }
 
 mod base64url {
+    use base64ct::{Base64UrlUnpadded, Encoding};
     use serde::{
         de::{Error as DeError, Unexpected, Visitor},
         Deserializer, Serializer,
@@ -494,7 +495,7 @@ mod base64url {
         S: Serializer,
     {
         if serializer.is_human_readable() {
-            serializer.serialize_str(&base64::encode_config(value, base64::URL_SAFE_NO_PAD))
+            serializer.serialize_str(&Base64UrlUnpadded::encode_string(value))
         } else {
             serializer.serialize_bytes(value)
         }
@@ -514,7 +515,7 @@ mod base64url {
             }
 
             fn visit_str<E: DeError>(self, value: &str) -> Result<Self::Value, E> {
-                base64::decode_config(value, base64::URL_SAFE_NO_PAD)
+                Base64UrlUnpadded::decode_vec(value)
                     .map_err(|_| E::invalid_value(Unexpected::Str(value), &self))
             }